+++ src/containers/group/groupId/groupTaskList/GroupPage.tsx
@@ -13,7 +13,6 @@ function GroupPage() {
   const { group, isGroupLoading, groupError, groupTaskLists, groupMembers } =
     useGroups(Number(groupId));
 
-
   // 전체 할 일, 전체 한 일, 오늘의 할 일
   const [doneTaskCount, setTaskDoneCount] = useState<number>(0);
   const [todayTaskCount, setTodayTaskCount] = useState<number>(0);
